+++ internal/driver/cluster/balancer/conn/conn.go
@@ -91,9 +91,11 @@ func (c *conn) take(ctx context.Context) (cc *grpc.ClientConn, err error) {
 	defer c.Unlock()
 	if isBroken(c.cc) {
 		_ = c.close(ctx)
-		var cancel context.CancelFunc
-		ctx, cancel = context.WithTimeout(ctx, c.config.DialTimeout())
-		defer cancel()
+		if dialTimeout := c.config.DialTimeout(); dialTimeout > 0 {
+			var cancel context.CancelFunc
+			ctx, cancel = context.WithTimeout(ctx, dialTimeout)
+			defer cancel()
+		}
 		cc, err = c.dial(ctx, c.endpoint.Address())
 		if err != nil {
 			return nil, err