+++ test/ratelimiter_test.go
@@ -1,16 +1,16 @@
-// +build integration
-
 package test
 
 import (
 	"context"
+	"os"
 	"testing"
 	"time"
 
-	ydb "github.com/ydb-platform/ydb-go-sdk/v3"
+	"github.com/ydb-platform/ydb-go-sdk/v3"
 	"github.com/ydb-platform/ydb-go-sdk/v3/config"
 	cfg "github.com/ydb-platform/ydb-go-sdk/v3/coordination"
 	"github.com/ydb-platform/ydb-go-sdk/v3/internal/coordination"
+	"github.com/ydb-platform/ydb-go-sdk/v3/internal/errors"
 	internal "github.com/ydb-platform/ydb-go-sdk/v3/internal/ratelimiter"
 	public "github.com/ydb-platform/ydb-go-sdk/v3/ratelimiter"
 )
@@ -21,13 +21,20 @@ const (
 )
 
 func TestRateLimiter(t *testing.T) {
-	t.Skip("run under docker")
+	if testing.Short() {
+		t.Skip("skipping testing in short mode")
+	}
+
 	ctx, cancel := context.WithTimeout(context.Background(), time.Second*2)
 	defer cancel()
 
-	db, err := open(
+	db, err := ydb.New(
 		ctx,
-		ydb.WithDriverConfigOptions(
+		ydb.WithConnectionString(os.Getenv("YDB_CONNECTION_STRING")),
+		ydb.WithAnonymousCredentials(),
+		ydb.WithTraceDriver(driverTrace()),
+		ydb.WithTraceTable(tableTrace()),
+		ydb.With(
 			config.WithRequestTimeout(time.Second*5),
 			config.WithStreamTimeout(time.Second*5),
 			config.WithOperationTimeout(time.Second*5),
@@ -46,7 +53,12 @@ func TestRateLimiter(t *testing.T) {
 
 	err = coordinationClient.DropNode(ctx, testCoordinationNodePath)
 	if err != nil {
-		t.Fatal(err)
+		if d := ydb.TransportErrorDescription(err); d != nil && d.Code() == int32(errors.TransportErrorUnimplemented) {
+			return
+		}
+		if d := ydb.OperationErrorDescription(err); d != nil && d.Code() != int32(errors.StatusSchemeError) {
+			t.Fatal(err)
+		}
 	}
 	err = coordinationClient.CreateNode(ctx, testCoordinationNodePath, cfg.Config{
 		Path:                     "",